chore(http): move root route
