+++ src/http-server/root.js
@@ -1,0 +1,30 @@
+// Import Node.js Dependencies
+import { join, dirname } from "path";
+import { readFile } from "fs/promises";
+import { fileURLToPath } from "url";
+
+// Import Third-party Dependencies
+import zup from "zup";
+import * as i18n from "@nodesecure/i18n";
+
+const __dirname = dirname(fileURLToPath(import.meta.url));
+const kProjectRootDir = join(__dirname, "..", "..");
+
+export async function root(req, res) {
+  try {
+    res.writeHead(200, {
+      "Content-Type": "text/html"
+    });
+
+    const HTMLStr = await readFile(join(kProjectRootDir, "views", "index.html"), "utf-8");
+    const templateStr = zup(HTMLStr)({
+      lang: i18n.getToken("lang"),
+      token: (tokenName) => i18n.getToken(`ui.${tokenName}`)
+    });
+
+    res.end(templateStr);
+  }
+  catch (err) {
+    send(res, 500, { error: err.message });
+  }
+}
